+++ packages/executor-react/CHANGELOG.md
@@ -1,4 +1,24 @@
-# @midl-xyz/executor-react
+# @midl-xyz/midl-js-executor-react
+
+## 0.0.4
+
+### Patch Changes
+
+- fix: types for add tx intention
+
+## 0.0.3
+
+### Patch Changes
+
+- fix: export hooks
+
+## 0.0.2
+
+### Patch Changes
+
+- feat: update contracts
+- Updated dependencies
+  - @midl-xyz/midl-js-executor@0.0.75
 
 ## 0.0.1